Fixed triangle hit function, working tetrahedral
Issue with triangle hit was with the mathematics behind the logic.
While using dot product with the edges of the triangle works to provide
the point's location within the parallelogram formed by the triangle if
the triangle is obtuse, it does not work if it is acute. This is
because a point outside the triangle can still have a dot product with
the sides of greater than 1, which would qualify it to be part of the
triangle.

To solve this, the logic was switched from using a vector coordinate system (which doesn't seem to work, even if 3 vectors are used) to an area-based system. The cross product of two vectors is equal to twice the magnitude of the area formed by the triangle between the two vectors. Thus, cross product is used to compute a vector for the main triangle, and area of triangles formed by edges with the point, which represent barycentric coordinates. These vectors are compared with normal vector (to which they should be parallel) to determine their size. If the signs of the vectors are the same, then they indicate that the triangles are all internal to the main triangle, and the point is also within the main triangle. If the signs are not the same, the point lies outside the main triangle.