Added intuitive tetrahedral constructor
A new constructor for tetrahedral has been made so that the 'top' vector argument now specifies direction of one of the points so that 'top is orthogonal to one of the bases. 'Front' now specifies the direction of another one of the faces, although front will be orthogonal to top (parallel to base).